+++ multi_label_classifier/preprocess/radardata_parser.py
@@ -7,7 +7,7 @@ import os
 import logging
 from datasets import load_dataset
 import pandas as pd
-from ..constants import RADAR_DIR, RADAR_RAW_DIR
+from ..settings import RADAR_DIR, RADAR_RAW_DIR
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
